fix bug where an unchecked checkbox, which is required, was getting the isNonEmptyString validator method added to it, thereby erroneously making it seem invalid
